SOLID
aplicar los principios SOLID en los
Controllers auth y product
