+++ BACKEND/controllers/authController/Logout.js
@@ -1,0 +1,3 @@
+exports.logout = (req, res) => {
+    res.json({ message: "logout exitoso" });
+  };
