+++ BACKEND/controllers/productController/getProduct.js
@@ -1,0 +1,18 @@
+const Product = require("../models/Product");
+
+exports.getProduct = async (req, res) => {
+  const { id } = req.params;
+
+  const product = await Product.findOne({
+    where: {
+      id: id,
+      isDeleted: false
+    }
+  });
+
+  if (!product) {
+    return res.status(404).json({ message: "Producto no encontrado o eliminado" });
+  }
+
+  res.json(product);
+};
