+++ BACKEND/controllers/productController/getProducts.js
@@ -1,0 +1,10 @@
+const Product = require("../models/Product");
+
+exports.getProducts = async (req, res) => {
+  const products = await Product.findAll({
+    where: {
+      isDeleted: false
+    }
+  });
+  res.json(products);
+};
